Former-commit-id: 3c1cdf6b2b8e129c1bef380f5b991af7e8aafd79

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/NumericIntegerQuestionAnswered.cs
@@ -11,7 +11,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
     {
         public int Answer { get; private set; }
 
-        public NumericIntegerQuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime, int answer)
+        public NumericIntegerQuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime, int answer)
             : base(userId, questionId, propagationVector, answerTime)
         {
             this.Answer = answer;
