+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/SingleOptionQuestionAnswered.cs
@@ -7,7 +7,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
     {
         public decimal SelectedValue { get; private set; }
 
-        public SingleOptionQuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime, decimal selectedValue)
+        public SingleOptionQuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime, decimal selectedValue)
             : base(userId, questionId, propagationVector, answerTime)
         {
             this.SelectedValue = selectedValue;
