+++ src/Tests/WB.Tests.Integration/InterviewTests/when_decreasing_propagation_count_and_latest_propagated_question_was_used_in_linked_single_option_question_as_selected_answer.cs
@@ -68,21 +68,21 @@ namespace WB.Tests.Integration.InterviewTests
             };
 
             interview = CreateInterviewFromQuestionnaireDocumentRegisteringAllNeededDependencies(questionnaireDocument);
-            interview.AnswerNumericIntegerQuestion(userId, numericQuestionId, new int[] {}, answerTime, 3);
-            interview.AnswerTextQuestion(userId, referencedQuestionId, new int[] { 0 }, answerTime, "A");
-            interview.AnswerTextQuestion(userId, referencedQuestionId, new int[] { 2 }, answerTime, "C");
-            interview.AnswerSingleOptionLinkedQuestion(userId, linkedQuestionId, new int[] {}, answerTime, new int[] { 2 });
+            interview.AnswerNumericIntegerQuestion(userId, numericQuestionId, new decimal[] { }, answerTime, 3);
+            interview.AnswerTextQuestion(userId, referencedQuestionId, new decimal[] { 0 }, answerTime, "A");
+            interview.AnswerTextQuestion(userId, referencedQuestionId, new decimal[] { 2 }, answerTime, "C");
+            interview.AnswerSingleOptionLinkedQuestion(userId, linkedQuestionId, new decimal[] { }, answerTime, new decimal[] { 2 });
 
             eventContext = new EventContext();
         };
 
         Because of = () =>
-            interview.AnswerNumericIntegerQuestion(userId, numericQuestionId, new int[] {}, answerTime, 2);
+            interview.AnswerNumericIntegerQuestion(userId, numericQuestionId, new decimal[] { }, answerTime, 2);
 
         It should_raise_AnswerRemoved_event_with_linked_single_option_question_id_and_propagation_vector = () =>
             eventContext.ShouldContainEvent<AnswerRemoved>(@event
                 => @event.QuestionId == linkedQuestionId
-                && @event.PropagationVector.SequenceEqual(new int[] {}));
+                && @event.PropagationVector.SequenceEqual(new decimal[] { }));
 
         Cleanup stuff = () =>
         {
